Create ZetkinCollapsible component and apply to PrettyEmail and JourneyInstanceSummary.
